game is mostly stable
except crosshair seems to start at an outer position after a reset of the game

+++ Inc/PuTTYSprites.h
@@ -8,7 +8,7 @@
 
 void ui_draw_sprite(uint8_t index, uint8_t FGC, uint8_t BGC, uint8_t x, uint8_t y);
 void ui_clear_sprite(uint8_t index, uint8_t FGC, uint8_t BGC, uint8_t x, uint8_t y);
-int8_t offsetBulletCoordX(player_t *player);
-int8_t offsetBulletCoordY(player_t *player);
+int16_t offsetBulletCoordX(player_t *player);
+int16_t offsetBulletCoordY(player_t *player);
 
 #endif /* PUTTYSPRITES_H_ */
